Refactor stylistically, add @Transactional where necessary
